Function to plot flux maps per time interval
This function is similar to 'plot_spatial_flux', however, it allows the plotting of only one variable (prior, posterior or difference).
Differently from 'plot_spatial_flux', the average is not performed over the full time window specified in start_date/end_date but considering the number of time stamps specified in 'dt'.
The function allows the plotting of one model or multiple models.
'dt' can be any integer number. If it is greater than the total number of time stamps, the number of time stamps is used instead.